Feature: Added WeaponButtonReactor script
This script will assign the proper Watcher depending on the hand (Left/Right) when an object is picked up using the XRInteractable event system.
The "PrimaryButtonPressed" and SecondaryButtonPressed events are also added to the Watcher upon pickup.
When the object is dropped, the events are removed and the Watcher is cleared.