Attach clip to weapon
Decided not to use XR Sockets for Clip attachement on the weapons.
Used regular OnTrigger methods
Mechanics done, can now attach a clip to the Pistol and Drop it when pressing PrimaryButton

Had issues with Unity's XRDirectInteractor, the interactor keeps checking the items around in order to interact with them, if you destroy a GameObject, the interactor does not do a NULL check.

Had to extend the class and fix the "GetValidTargets" method