Fixed Magazine dropping out of gun
Fixed issue where magazine was dropping from gun when attached with hand (was working correctly when you would touch the mag on the table with the gun)

The underlying issue was related to the Hand XRGrabInteractable/Interactor.

When attaching the mag, the hand would keep the item as “selected” even though the parent was reassigned.

As I was destroying the XRGrabInteractable when attaching the magazine, to prevent the player from grabbing the magazine back directly from the gun without releasing it, it was causing the item to completely detach and fall through the gun and floor.

Resolved issue by Instantiating a new magazine to replace original one and destroy the original magazine immediately, resulting in the clone not being “held” and staying in place