Reloading Mechanics
In order to avoid issues with trigger/collision when you drop the magazine, the magazine will be "unloadable" for 0.5 seconds to allow it to exit the weapon

If its not empty, you will be able to reload your gun again with it after that delay

Added "AmmoContainer" script to AmmoBox prefabs

Can now define if ammo box/magazine is a magazine (IsMagazine) - This is used to destroy the item correctly when there is no ammo in it.
If IsMagazine, it will wait for it to be dropped out of the gun before destroying it when empty, if not (AmmoBox), it will simply destroy it when it is empty.

Refined Shotgun and Rifle loading mechanic. When triggering the ammobox on the weapon, it will check the ammo the weapon needs to be full and transfer the correct ammo count from the box to the weapon.

Minor tweaks in firing/reloading mechanics:

If the weapon has a magazine, the magazine will be depleted instead of the weapon. Before, it was adding the bullets to the weapon and transferring back the remaining bullets into the magazine when dropped.

DUK-17