Update on billboard and game instance reset
DUK-49 The billboard now display the proper values even when the game instance is reset -> which it previously didnt.
The createNewGame class was transform and is no longer a singleton as it was causing trouble to reset its value. It is now a simple class that is part of the Scoring System