Fixed outline not appearing after dropping magazine
Had issues when the magazine was dropped out of a weapon.

Due to the way the reloading mechanic works, the events that would trigger the Selection outline were destroyed when magazine was inserted.

This caused the outline not showing after a magazine was inserted and dropped from a weapon.

Fixed issue by adding back events through code on magazine drop

DUK-34