AmmoCrates fixed and tweaks
AmmoCrates now spawn ammo on a timer once it has been emptied.

To also prevent player from "loosing" an ammo magazine/box and break the AmmoCrates spawning mechanics, the ammo box/magazine will now be destroyed automatically if the distance between the AmmoCrate and the ammo box/magazine is to long.

DUK-44