Fixes and tweaks on Ammo Crates
When a player now picks up an ammo magazine/box, it will not have a parent anymore so the crate will immediately remove it from its content.

Also, if the ammo box is taken out of the box without picking it up (knocked out of the box by an object for example) it will also be removed from the box when the distance between the crate and magazine/box is too high.

DUK-44