Weapon reset mechanic
Created a "WeaponSpawn" script that can be attached to a gameobject and it will instantiate the weapon that is linked in the inspector

Created 4 "WeaponSpawner" prefabs (PistolSpawn, Submachinegun, Shotgun, Rifle)
When the game starts, the "WeaponResetButton" script will be loaded first in the Script Execution order.

After that, each "WeaponSpawner" will register themselves with the WeaponResetButton and the weapon reset button will Instanciate every weapons on Start.

If the button is clicked, it will destroy all the weapons and instantiate them again.

DUK-56